Ajout du ZoomCommand et initialisation du imageModel

+++ src/main/java/com/example/laboratoire_5/Perspective.java
@@ -23,7 +23,7 @@ public class Perspective {
         return scale;
     }
 
-    public void setScale(float scale) {
+    public void setScale(double scale) {
         this.scale = scale;
     }
